update backend for mcp
