+++ backend/src/api/middleware/auth.ts
@@ -2,7 +2,7 @@ import { Request, Response, NextFunction } from 'express';
 import { AuthService } from '@/core/auth/auth.js';
 import { AppError } from './error.js';
 import { ERROR_CODES, NEXT_ACTION } from '@/types/error-constants.js';
-import { createRemoteJWKSet, jwtVerify } from 'jose';
+import { verifyCloudToken } from '@/utils/cloud-token.js';
 
 export interface AuthRequest extends Request {
   user?: {
@@ -15,15 +15,7 @@ export interface AuthRequest extends Request {
   projectId?: string;
 }
 
-// Lazy initialization to avoid circular dependency
-let authService: AuthService | null = null;
-
-function getAuthService(): AuthService {
-  if (!authService) {
-    authService = AuthService.getInstance();
-  }
-  return authService;
-}
+const authService = AuthService.getInstance();
 
 // Helper function to extract Bearer token
 function extractBearerToken(authHeader: string | undefined): string | null {
@@ -42,38 +34,6 @@ function setRequestUser(req: AuthRequest, payload: { sub: string; email: string;
   };
 }
 
-// Helper function to verify cloud backend JWT token
-export async function verifyCloudToken(token: string): Promise<{ projectId: string; payload: any }> {
-  // Create JWKS endpoint for remote key set
-  const JWKS = createRemoteJWKSet(
-    new URL(
-      (process.env.CLOUD_API_HOST || 'https://api.insforge.dev') + '/.well-known/jwks.json'
-    )
-  );
-
-  // Verify the token with jose
-  const { payload } = await jwtVerify(token, JWKS, {
-    algorithms: ['RS256', 'RS384', 'RS512', 'ES256', 'ES384', 'ES512'],
-  });
-
-  // Verify project_id matches if configured
-  const tokenProjectId = (payload as any).projectId;
-  const expectedProjectId = process.env.PROJECT_ID;
-  
-  if (expectedProjectId && tokenProjectId !== expectedProjectId) {
-    throw new AppError(
-      'Project ID mismatch',
-      403,
-      ERROR_CODES.AUTH_UNAUTHORIZED,
-      NEXT_ACTION.CHECK_TOKEN
-    );
-  }
-
-  return {
-    projectId: tokenProjectId,
-    payload
-  };
-}
 
 /**
  * Verifies user authentication (accepts both user and admin tokens)
@@ -111,7 +71,7 @@ export async function verifyAdmin(req: AuthRequest, res: Response, next: NextFun
     }
 
     // For admin, we use JWT tokens
-    const payload = getAuthService().verifyToken(token);
+    const payload = authService.verifyToken(token);
 
     if (payload.role !== 'project_admin') {
       throw new AppError(
@@ -152,7 +112,7 @@ export async function verifyApiKey(req: AuthRequest, _res: Response, next: NextF
       );
     }
 
-    const isValid = await getAuthService().verifyApiKey(apiKey);
+    const isValid = await authService.verifyApiKey(apiKey);
     if (!isValid) {
       throw new AppError(
         'Invalid API key',
@@ -187,7 +147,7 @@ export async function verifyToken(req: AuthRequest, _res: Response, next: NextFu
     }
 
     // Verify JWT token
-    const payload = getAuthService().verifyToken(token);
+    const payload = authService.verifyToken(token);
 
     // Validate token has a role
     if (!payload.role) {
